requirements.txt
# Core Framework
fastapi>=0.128.5
uvicorn[standard]>=0.30.0
pydantic>=2.12.5
pydantic-settings>=2.6.0

# AI & LLM - LangChain v1
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Google Gemini
google-generativeai>=0.8.0

# Memory & RAG
mem0ai>=1.0.3
chromadb>=1.4.1

# Slack Integration
slack-bolt>=1.27.0
slack-sdk>=3.39.0

# Database
sqlalchemy>=2.0.46

# Task Scheduling
apscheduler>=3.11.2

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0

# Google APIs for Email & Calendar
google-api-python-client==2.117.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
